\name{emudata-package}
\alias{emudata-package}
\alias{emudata}
\docType{package}
\title{
Datasets for the emu package}
\description{
The datasets contain examples for some functions.
Following datasets included:
  dorfric
  dorsal
  f2geraus
  fr
  keng
  plos
  sib
  stops10
  stops2
  timevow
  v.sam
}
\details{
\tabular{ll}{
Package: \tab emu-data\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-07-03\cr
License: \tab Copyright 2012 IPS LMU Munich. All rights reserved. \cr
}
Datasets load automatically. 
}
\author{
Tina John (help text)

Maintainer: <jmh@phonetik.uni-muenchen.de>
}
\references{
  Harrington, J. (2010). The Phonetic Analysis of Speech Corpora. Blackwell.
}
\keyword{ package }
\seealso{
\code{\link[emu:emu-package]{emu.testsuite}}
}
\examples{
  data(fr)
}
